pyproject.toml
[project]
name = "knowledge-distillation-basic"
version = "0.1.0"
description = ""
authors = [
    {name = "Le Hoang",email = "lehoangnguyen510@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib (>=3.10.1,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "torchvision (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
